+++ qatrack/qatrack_core/serializers.py
@@ -49,13 +49,13 @@ class QATrackJSONEncoder(DjangoJSONEncoder):
                 return method()
 
         if isinstance(o, datetime.datetime):
-            r = o.strftime(settings.DATETIME_INPUT_FORMATS[1])
+            r = o.strftime(settings.QATRACK_API_DATETIME_FORMAT)
             if o.microsecond:
                 r = r[:23] + r[26:]
             if r.endswith('+00:00'):
                 r = r[:-6] + 'Z'
             return r
         elif isinstance(o, datetime.date):
-            return o.strftime(settings.DATE_INPUT_FORMATS[0])
+            return o.strftime(settings.QATRACK_API_DATE_FORMAT)
 
         return super().default(o)