+++ qatrack/reports/tests/test_base.py
@@ -1,7 +1,6 @@
 import datetime
 import io
 import json
-import time
 from unittest import mock
 
 from django.contrib.admin.sites import AdminSite
@@ -13,13 +12,10 @@ from django.test.utils import override_settings
 from django.urls import reverse
 from django.utils import timezone
 from django_q.models import Schedule
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as e_c
 
 # Create your tests here.
 from qatrack.qa.models import Group, TestInstance, UnitTestCollection, User
 from qatrack.qa.tests import utils
-from qatrack.qa.tests.test_selenium import BaseQATests
 from qatrack.reports import (
     admin,
     filters,
@@ -139,7 +135,7 @@ class TestReportPreview(TestCase):
             'root-title': 'Title',
             'root-report_format': 'pdf',
             'root-include_signature': True,
-            'work_completed': '01 Jan 2000',
+            'work_completed': '2000-01-01',
             'reportnote_set-INITIAL_FORMS': 0,
             'reportnote_set-TOTAL_FORMS': 0,
         }
@@ -196,7 +192,7 @@ class TestSaveReport(TestCase):
             'root-title': 'Title',
             'root-report_format': 'pdf',
             'root-include_signature': True,
-            'work_completed': '01 Jan 2000',
+            'work_completed': '2000-01-01',
             'reportnote_set-INITIAL_FORMS': 0,
             'reportnote_set-TOTAL_FORMS': 0,
         }
@@ -654,7 +650,8 @@ class TestInstanceToFormFields(TestCase):
         sr = models.SavedReport(report_type=qc.TestListInstanceSummaryReport.report_type)
         sr.filters = {'work_completed': ["1 Jan 2019", "2 Jan 2019"]}
         res = forms.serialize_savedreport(sr)
-        assert res['work_completed'] == ['text', '01 Jan 2019 - 02 Jan 2019']
+        # typed in one format, shown back in the configured default
+        assert res['work_completed'] == ['text', '2019-01-01 - 2019-01-02']
 
 
 class TestReportToFormFields(TestCase):
@@ -666,7 +663,8 @@ class TestReportToFormFields(TestCase):
     def test_daterange(self):
         report = qc.TestListInstanceSummaryReport(report_opts={'work_completed': ["1 Jan 2019", "2 Jan 2019"]})
         res = forms.serialize_report(report)
-        assert res['work_completed'] == ['text', '01 Jan 2019 - 02 Jan 2019']
+        # typed in one format, shown back in the configured default
+        assert res['work_completed'] == ['text', '2019-01-01 - 2019-01-02']
 
     def test_visible_to(self):
         g = Group.objects.create(name="group")
@@ -783,24 +781,24 @@ class TestBaseReport(TestCase):
     @override_settings(TIME_ZONE="America/Toronto")
     def test_default_detail_value_format_datetime_utc(self):
         dt = timezone.datetime(2019, 1, 2, 2, 0, tzinfo=datetime.UTC)
-        assert reports.BaseReport().default_detail_value_format(dt) == "01 Jan 2019"
+        assert reports.BaseReport().default_detail_value_format(dt) == "2019-01-01"
 
     @override_settings(TIME_ZONE="America/Toronto")
     def test_default_detail_value_format_datetime_naive(self):
         dt = timezone.datetime(2019, 1, 2, 2, 0)
-        assert reports.BaseReport().default_detail_value_format(dt) == "02 Jan 2019"
+        assert reports.BaseReport().default_detail_value_format(dt) == "2019-01-02"
 
     @override_settings(TIME_ZONE="America/Toronto")
     def test_default_detail_value_format_datetime_range_utc(self):
         dt1 = timezone.datetime(2019, 1, 2, 2, 0, tzinfo=datetime.UTC)
         dt2 = timezone.datetime(2019, 1, 3, 2, 0, tzinfo=datetime.UTC)
-        assert reports.BaseReport().default_detail_value_format([dt1, dt2]) == "01 Jan 2019 - 02 Jan 2019"
+        assert reports.BaseReport().default_detail_value_format([dt1, dt2]) == "2019-01-01 - 2019-01-02"
 
     @override_settings(TIME_ZONE="America/Toronto")
     def test_default_detail_value_format_datetime_range_naive(self):
         dt1 = timezone.datetime(2019, 1, 2, 2, 0)
         dt2 = timezone.datetime(2019, 1, 3, 2, 0)
-        assert reports.BaseReport().default_detail_value_format([dt1, dt2]) == "02 Jan 2019 - 03 Jan 2019"
+        assert reports.BaseReport().default_detail_value_format([dt1, dt2]) == "2019-01-02 - 2019-01-03"
 
     def test_default_detail_value_format_iterable(self):
         assert reports.BaseReport().default_detail_value_format([1, 2]) == "1, 2"
@@ -842,238 +840,6 @@ class TestBaseReport(TestCase):
         xls = rep.to_xlsx()
         assert isinstance(xls, io.BytesIO)
         assert xls.tell() == 0
-
-
-class TestReportInterface(BaseQATests):
-
-    def setUp(self):
-        super().setUp()
-        self.login()
-        self.open(reverse("reports"))
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'select2-id_root-report_type-container')))
-
-    def test_report_preview(self):
-        """Select report and make sure it previews"""
-        self.select_by_text('id_root-report_type', qc.TestListInstanceSummaryReport.name)
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'id_work_completed')))
-        self.click("preview")
-        self.driver.find_element(By.CSS_SELECTOR, '#report .container-fluid')
-
-    def test_save_report(self):
-        """Ensure filling and saving a report results in a SavedReport in the db"""
-        self.select_by_text('id_root-report_type', qc.TestListInstanceSummaryReport.name)
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'id_work_completed')))
-        assert models.SavedReport.objects.count() == 0
-        self.click("save")
-        self.wait.until(e_c.presence_of_element_located((By.CLASS_NAME, 'success-message')))
-        assert models.SavedReport.objects.count() == 1
-        sr = models.SavedReport.objects.first()
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'report-id-%s' % sr.pk)))
-
-    def test_save_report_with_note(self):
-        """Ensure adding notes to saved reports works"""
-        self.select_by_text('id_root-report_type', qc.TestListInstanceSummaryReport.name)
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'id_work_completed')))
-        self.click("add-note")
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'id_reportnote_set-0-heading')))
-        self.send_keys("id_reportnote_set-0-heading", "heading")
-        self.send_keys("id_reportnote_set-0-content", "content")
-
-        assert models.ReportNote.objects.count() == 0
-        self.click("save")
-        self.wait.until(e_c.presence_of_element_located((By.CLASS_NAME, 'success-message')))
-        expected_notes = [{"heading": "heading", "content": "content"}]
-        assert list(models.ReportNote.objects.values("heading", "content")) == expected_notes
-
-    def test_save_report_with_note_repeated_saves(self):
-        """Ensure repeated saves only create one note"""
-        self.select_by_text('id_root-report_type', qc.TestListInstanceSummaryReport.name)
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'id_work_completed')))
-        self.click("add-note")
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'id_reportnote_set-0-heading')))
-        self.send_keys("id_reportnote_set-0-heading", "heading")
-        self.send_keys("id_reportnote_set-0-content", "content")
-
-        assert models.ReportNote.objects.count() == 0
-        for i in range(3):
-            self.click("save")
-            self.wait.until(e_c.presence_of_element_located((By.CLASS_NAME, 'success-message')))
-        expected_notes = [{"heading": "heading", "content": "content"}]
-        assert list(models.ReportNote.objects.values("heading", "content")) == expected_notes
-
-    def test_load_report(self):
-        """Select report from table and make sure it loads"""
-
-        sr = models.SavedReport.objects.create(
-            report_type=qc.TestListInstanceSummaryReport.report_type,
-            report_format="pdf",
-            title="title",
-            filters={'work_completed': ['2 Jan 1989', '4 Jan 1990']},
-            created_by=self.user,
-            modified_by=self.user,
-        )
-        models.ReportNote.objects.create(
-            report=sr,
-            heading="heading",
-            content="content",
-        )
-
-        # need to reload page to get report table
-        self.driver.refresh()
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'report-id-%s' % sr.pk)))
-        self.click('report-id-%s' % sr.pk)
-        wc = self.driver.find_element(By.ID, 'id_work_completed')
-        assert wc.get_attribute("value") == "02 Jan 1989 - 04 Jan 1990"
-        heading = self.driver.find_element(By.ID, "id_reportnote_set-0-heading")
-        assert heading.get_attribute("value") == "heading"
-        content = self.driver.find_element(By.ID, "id_reportnote_set-0-content")
-        assert content.get_attribute("value") == "content"
-
-    def test_load_report_edit_note(self):
-        """Select report from table, edit its note and resave it"""
-
-        sr = models.SavedReport.objects.create(
-            report_type=qc.TestListInstanceSummaryReport.report_type,
-            report_format="pdf",
-            title="title",
-            filters={'work_completed': ['2 Jan 1989', '4 Jan 1990']},
-            created_by=self.user,
-            modified_by=self.user,
-        )
-        models.ReportNote.objects.create(
-            report=sr,
-            heading="heading",
-            content="content",
-        )
-
-        # need to reload page to get report table
-        self.driver.refresh()
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'report-id-%s' % sr.pk)))
-        self.click('report-id-%s' % sr.pk)
-        heading = self.driver.find_element(By.ID, "id_reportnote_set-0-heading")
-        heading.send_keys(" add some new text")
-        self.click("save")
-        self.wait.until(e_c.presence_of_element_located((By.CLASS_NAME, 'success-message')))
-        expected_notes = [{"heading": "heading add some new text", "content": "content"}]
-        assert list(models.ReportNote.objects.values("heading", "content")) == expected_notes
-
-    def test_load_report_delete_note(self):
-        """Select report from table, delete a note and then save it """
-
-        sr = models.SavedReport.objects.create(
-            report_type=qc.TestListInstanceSummaryReport.report_type,
-            report_format="pdf",
-            title="title",
-            filters={'work_completed': ['2 Jan 1989', '4 Jan 1990']},
-            created_by=self.user,
-            modified_by=self.user,
-        )
-        models.ReportNote.objects.create(
-            report=sr,
-            heading="heading",
-            content="content",
-        )
-
-        # need to reload page to get report table
-        self.driver.refresh()
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'report-id-%s' % sr.pk)))
-        self.click('report-id-%s' % sr.pk)
-        self.click("id_reportnote_set-remove-0")
-        self.click("save")
-        self.wait.until(e_c.presence_of_element_located((By.CLASS_NAME, 'success-message')))
-        assert models.ReportNote.objects.count() == 0
-
-    def test_load_report_add_new_note_delete_old_note(self):
-        """Ensure we can both add and delete notes in a single save"""
-
-        sr = models.SavedReport.objects.create(
-            report_type=qc.TestListInstanceSummaryReport.report_type,
-            report_format="pdf",
-            title="title",
-            filters={'work_completed': ['2 Jan 1989', '4 Jan 1990']},
-            created_by=self.user,
-            modified_by=self.user,
-        )
-        models.ReportNote.objects.create(
-            report=sr,
-            heading="heading",
-            content="content",
-        )
-
-        # need to reload page to get report table
-        self.driver.refresh()
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'report-id-%s' % sr.pk)))
-        self.click('report-id-%s' % sr.pk)
-        self.click("add-note")
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'id_reportnote_set-1-heading')))
-        self.send_keys("id_reportnote_set-1-heading", "heading new")
-        self.send_keys("id_reportnote_set-1-content", "content new")
-        self.click("id_reportnote_set-remove-0")
-        self.click("save")
-        self.wait.until(e_c.presence_of_element_located((By.CLASS_NAME, 'success-message')))
-        expected_notes = [{"heading": "heading new", "content": "content new"}]
-        assert list(models.ReportNote.objects.values("heading", "content")) == expected_notes
-
-    def test_schedule_report(self):
-        """Ensure scheduling a savedreport works"""
-
-        sr = models.SavedReport.objects.create(
-            report_type=qc.TestListInstanceSummaryReport.report_type,
-            report_format="pdf",
-            title="title",
-            filters={'work_completed': ['2 Jan 1989', '4 Jan 1990']},
-            created_by=self.user,
-            modified_by=self.user,
-        )
-        # need to reload page to get report table
-        self.driver.refresh()
-
-        self.click('report-id-%s' % sr.pk)
-
-        self.click('report-id-%s-schedule' % sr.pk)
-
-        self.select_by_index('id_schedule-time', 1)
-        self.driver.find_element(By.ID, "id_schedule-emails").send_keys("a@b.com")
-
-        self.wait.until(e_c.presence_of_element_located((By.CLASS_NAME, 'add-date')))
-        self.driver.find_element(By.CLASS_NAME, "add-date").click()
-
-        self.click("schedule")
-
-        self.wait.until(e_c.presence_of_element_located((By.CLASS_NAME, 'alert-success')))
-        sched = str(models.ReportSchedule.objects.first().schedule)
-        assert timezone.localtime(timezone.now()).strftime("%Y%m%d") in sched
-
-    def test_clear_schedule(self):
-        """Test clearing the schedule from a saved report"""
-
-        sr = models.SavedReport.objects.create(
-            report_type=qc.TestListInstanceSummaryReport.report_type,
-            report_format="pdf",
-            title="title",
-            filters={'work_completed': ['2 Jan 1989', '4 Jan 1990']},
-            created_by=self.user,
-            modified_by=self.user,
-        )
-        rec = "RRULE:FREQ=WEEKLY;BYDAY=MO,TU,WE"
-        models.ReportSchedule.objects.create(
-            report=sr,
-            time="00:00:00",
-            schedule=rec,
-            created_by=self.user,
-            modified_by=self.user,
-        )
-
-        # need to reload page to get report table
-        self.driver.refresh()
-        self.wait.until(e_c.presence_of_element_located((By.ID, 'report-id-%s' % sr.pk)))
-
-        self.click("report-id-%s-schedule" % sr.pk)
-        time.sleep(1)
-
-        self.click("clear-schedule")
-        self.wait.until(e_c.presence_of_element_located((By.CLASS_NAME, 'alert-success')))
-        assert models.ReportSchedule.objects.count() == 0
 
 
 class TestSavedReportAdmin: