+++ qatrack/units/tests/__init__.py
@@ -1,0 +1,1 @@
+from qatrack.units.tests.test_units import *  # NOQA